Expand `GET /health`'s test to cover the new value. Add new database methods for setting cluster-wide configuration values.
